% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePriorityMap.R
\name{makePriorityMap}
\alias{makePriorityMap}
\title{\code{makePriorityMapFN} makePriorityMap}
\usage{
makePriorityMap(useNewerData = TRUE, version = "2",
  bufferNests = 400, speciesOrder = NULL,
  interpolationDistance = 250, wholeIsland = FALSE)
}
\arguments{
\item{useNewerData}{Default is TRUE. Updates Dias et al., 2017 data with Sobral et al., in prep.}

\item{version}{Default is \code{"2"}. Version "1" uses slope as a proxy to marine birds' nests and does not
interpolate. Version "2" uses just the marine bird nests data and the cat density, both normalized.
It also interpolates the final raster to smooth the transition between zones.}

\item{bufferNests}{Default is set to 400. This is the distance in meters from the nests (used if version = "2"/"3")
that should be considered priority area. Takes into consideration cat's home range and area use.}

\item{speciesOrder}{Default is set to \code{NULL}. This parameter can be passed as the object generated by using the
function  \code{rankSpecies}. If not passed, the ranking is automatically created based on expert
opinion.}

\item{interpolationDistance}{Default is 250. This indicated how spread out the effect of nests should have on overall
priority.}

\item{wholeIsland}{Default is FALSE. If \code{TRUE} the risk map is done for the whole main island in FN. If
\code{FALSE} it is done only for the PARNAMAR and surroundings.}
}
\description{
This functions creates the priority map for cat management in
             Fernando de Noronha. It downloads all needed data and creates
             a raster object and saves it in the home directory
}
\examples{
\dontrun{
mapa <- makePriorityMap()
}
}
\author{
Tati Micheletti (TRIADE)
}
