% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normali.R
\name{normali}
\alias{normali}
\title{\code{makePriorityMapFN} normali}
\usage{
normali(ras)
}
\arguments{
\item{ras}{An \code{RasterLayer} object to be normalized.}
}
\description{
This functions normalizes relative rasters for better plotting
}
\examples{
library("raster")
ras <- raster::raster(ncol = 10, nrow = 10, val = runif(100, 0, 10))
rasNorm <- normali(ras)
}
