% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankSpecies.R
\name{rankSpecies}
\alias{rankSpecies}
\title{\code{rankSpecies} rankSpecies}
\usage{
rankSpecies(species = NULL, ranking = NULL)
}
\arguments{
\item{species}{Character vector giving the 11 species to assign priorities.}

\item{ranking}{Numeric. Should give priorities for the 11 species passed in \code{species}
argument. Has to be of length 11L. Pririty 1 is the lowest, the others are
relative to the maximum priority given.#'}
}
\description{
This functions creates the priority values for a bird species list.
             The default uses values provided by CEMAVE/ICMBio specialists.
             The list HAS to have all 11 species found in Fernando de Noronha and
             you have to provide 11 values for the relative importance of these,
             with \code{1} being the lowest priority.
}
\examples{
\dontrun{
spRankDefault <- rankSpecies()
spRank <- rankSpecies(species = c("Phaethon lepturus", "Phaethon aethereus", "Sula dactylatra",
                      "Sula sula", "Sula leucogaster", "Arenaria interpres", "Onychoprion fuscatus",
                      "Anous minutus", "Gygis alba", "Anous stolidus",
                      "Bubulcus ibis"),
                      ranking = seq(11:1))
}
}
\author{
Tati Micheletti (TRIADE)
}
