% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePriorityMap.R
\name{makePriorityMap}
\alias{makePriorityMap}
\title{\code{makePriorityMapFN} makePriorityMap}
\usage{
makePriorityMap()
}
\description{
This functions creates the priority map for cat management in
Fernando de Noronha. It downloads all needed data and creates
a raster object and saves it in the home directory
}
\examples{
\dontrun{
mapa <- makePriorityMap()
}
}
